Feat: add firebase storage emulator and refactor existing storage code (#109)
This PR adds the usage of the Firebase storage emulator for the tests. It also refactors all the code and tests that were using the the Firebase storage by putting all the storage code in a single class.
Some functions and classes were also removed/split in order to be able to test them efficiently and to reuse them in other classes.