Create notifications architecture for competitions and sightseeings (#131)
This PR adds all the architecture of the notifications for the competitions and sightseeings. It also improves the former work done on notifications by adding the creation of channels and by adding new tests.

In this PR all the remaining tests that were still using links from Firebase Storage were refactored to use images on Google Drive or Wikipedia instead. In fact, now the Default profile picture is hosted on the Google Drive of CultureQuest instead of Firestore Storage which will allow to save some upload and download band pass.