create all the Firebase methods and services for notifications (#140)
This PR provides the service that handles the reception of messages from Firebase Cloud Messaging to then trigger the correct notifications in the device of the user. Each notification is sent by one sender device to all the devices of the receiver by using the REST API of Firebase Cloud Messaging.

This PR also adds all the Firebase Database methods to set and get the device tokens of each user, that are used to send the notification, along with the methods to set and get the notifications of a user in the database.