Improve notifications behavior and make the code more robust (#154)
This PR customizes the intents that each type of notification starts when the user clicks on it. It works whether the app is in foreground, background or killed. Several bugs in the code had to be corrected to make this work when the app was killed. In fact the 'activeProfile" becomes "null" when the App is killed so the code was optimized in this PR in order to catch these cases and not crash the App.

Another issue is solved in this PR, in fact when displaying our own profile in the "DisplayProfileActivity" the follow button was still visible and clickable, so this PR removes this.

Finally, this PR also corrects a navigation bug that was triggered when clicking on the home icon in "DisplayProfileActivity", now it redirects to the Home fragment and not to the Scan fragment as previously.