Fix and improve seed synchronization among different apps (#167)
In the past tournament backend PR, the seed (used for pseudorandom tasks) was determined by the current week's number and year. This didn't correctly take into account that a tournament could occur across two different official weeks (delimited by the end of Sunday), so with two associated week numbers, leading to seed desynchronization between a user performing tournament-related actions (generating date, getting UID, generating or fetching it) on, say Saturday, and another on the next Monday.

This PR makes it so that seed doesn't rely on week numbers anymore, but is rather generated by the first user opening the app when it's time to schedule the tournament (either because a tournament has never been scheduled, or because the previous one is over).
Once the seed is generated, other users would fetch it, and all get synchronized this way.

This ensures that the seed is consistent for all users across the whole time frame from when the tournament has to get scheduled, to its termination when 7 days have passed from its start.